requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utils
python-dateutil==2.8.2
python-slugify==8.0.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
python-dotenv==1.0.1
